+++ sources/include/RTFEM/Math/MatrixMath.h
@@ -17,14 +17,14 @@ public:
      * @param matrix
      * @return
      */
-    Float ComputeDeterminant2(const Matrix& matrix);
+    double ComputeDeterminant2(const Matrix& matrix);
 
     /**
      * Computes Determinant of Square matrix
      * @param matrix
      * @return
      */
-    Float ComputeDeterminant(const Matrix& matrix);
+    double ComputeDeterminant(const Matrix& matrix);
 
     /**
      * Removes specified row and column thus new matrix has dimension [N-1 x M-1]
@@ -33,7 +33,7 @@ public:
      * @param column
      * @return
      */
-    Matrix ContractMatrix(const Matrix &matrix, UInt row, UInt column);
+    Matrix ContractMatrix(const Matrix &matrix, unsigned int row, unsigned int column);
 
     Matrix Transpose(const Matrix& matrix);
 };
