+++ sources/src/RTFEM/DataStructure/Matrix.cpp
@@ -4,29 +4,29 @@
 
 namespace rtfem {
 
-Matrix::Matrix(UInt row_count, UInt column_count) :
+Matrix::Matrix(unsigned int row_count, unsigned int column_count) :
         dimensions_{row_count, column_count}{
     InitData(0.0);
 }
 
-Matrix::Matrix(Float value, UInt row_count, UInt column_count) :
+Matrix::Matrix(double value, unsigned int row_count, unsigned int column_count) :
         dimensions_{row_count, column_count}{
     InitData(value);
 }
 
-Matrix::Matrix(std::initializer_list<std::initializer_list<Float>> lists) :
+Matrix::Matrix(std::initializer_list<std::initializer_list<double>> lists) :
         dimensions_{0,0} {
-    UInt max_size = 0;
+    unsigned int max_size = 0;
     for(const auto& list : lists){
         if(max_size < list.size())
             max_size = list.size();
     }
-    dimensions_ = MatrixDimension{(UInt)lists.size(), max_size};
+    dimensions_ = MatrixDimension{(unsigned int)lists.size(), max_size};
     InitData(0.0);
-    UInt i = 0;
+    unsigned int i = 0;
 
     for(const auto& list : lists){
-        UInt j = 0;
+        unsigned int j = 0;
         for(const auto& value : list){
             data_[i][j] = value;
             j++;
@@ -42,34 +42,34 @@ Matrix::Matrix(const MatrixDimension &&matrix_dimension) :
 
 Matrix::~Matrix() {}
 
-std::vector<Float>& Matrix::operator[](UInt i) {
+std::vector<double>& Matrix::operator[](unsigned int i) {
     return GetRow(i);
 }
 
-const std::vector<Float>& Matrix::operator[] (UInt i) const {
+const std::vector<double>& Matrix::operator[] (unsigned int i) const {
     return GetRow(i);
 }
 
-std::vector<Float>& Matrix::GetRow(UInt i){
+std::vector<double>& Matrix::GetRow(unsigned int i){
     if(i >= dimensions_.row_count)
         throw std::out_of_range("Matrix Index out of Range");
 
     return data_[i];
 }
 
-const std::vector<Float>& Matrix::GetRow(UInt i) const {
+const std::vector<double>& Matrix::GetRow(unsigned int i) const {
     if(i >= dimensions_.row_count)
         throw std::out_of_range("Matrix Index out of Range");
 
     return data_[i];
 }
 
-void Matrix::InitData(Float value){
+void Matrix::InitData(double value){
     data_.resize(dimensions_.row_count);
 
-    for(UInt i = 0; i < dimensions_.row_count; i++){
+    for(unsigned int i = 0; i < dimensions_.row_count; i++){
         data_[i].resize(dimensions_.column_count);
-        for(UInt j = 0; j < dimensions_.column_count; j++){
+        for(unsigned int j = 0; j < dimensions_.column_count; j++){
             data_[i][j] = 0;
         }
     }
@@ -80,10 +80,10 @@ bool Matrix::operator==(const Matrix &rhs) const {
             dimensions().row_count != rhs.dimensions().row_count)
         return false;
 
-    UInt n = dimensions().row_count;
-    UInt m = dimensions().column_count;
-    for(UInt i = 0; i < n; i++){
-        for(UInt j = 0; j < m; j++){
+    unsigned int n = dimensions().row_count;
+    unsigned int m = dimensions().column_count;
+    for(unsigned int i = 0; i < n; i++){
+        for(unsigned int j = 0; j < m; j++){
             if((*this)[i][j] != rhs[i][j])
                 return false;
         }
@@ -100,8 +100,8 @@ Matrix& Matrix::operator+=(const Matrix& rhs){
             this->dimensions().column_count != rhs.dimensions().column_count)
         throw std::invalid_argument("Matrix Addition: Dimensions are not valid");
 
-    for(UInt i = 0; i < this->dimensions().row_count; i++){
-        for(UInt j = 0; j < this->dimensions().column_count; j++){
+    for(unsigned int i = 0; i < this->dimensions().row_count; i++){
+        for(unsigned int j = 0; j < this->dimensions().column_count; j++){
             (*this)[i][j] = (*this)[i][j] + rhs[i][j];
         }
     }
@@ -110,8 +110,8 @@ Matrix& Matrix::operator+=(const Matrix& rhs){
 
 std::ostream& operator<<(std::ostream& os, const Matrix& m){
     const std::string seperator = "   ";
-    for(UInt i = 0; i < m.dimensions().row_count; i++){
-        for(UInt j = 0; j < m.dimensions().column_count; j++){
+    for(unsigned int i = 0; i < m.dimensions().row_count; i++){
+        for(unsigned int j = 0; j < m.dimensions().column_count; j++){
             os << std::to_string(m[i][j]) << seperator;
         }
         os << std::endl;
@@ -119,33 +119,33 @@ std::ostream& operator<<(std::ostream& os, const Matrix& m){
     return os;
 }
 
-Matrix operator*(Float x, const Matrix &m) {
+Matrix operator*(double x, const Matrix &m) {
     Matrix new_matrix(std::move(m.dimensions()));
 
-    for(UInt i = 0; i < m.dimensions().row_count; i++){
-        for(UInt j = 0; j < m.dimensions().column_count; j++){
+    for(unsigned int i = 0; i < m.dimensions().row_count; i++){
+        for(unsigned int j = 0; j < m.dimensions().column_count; j++){
             new_matrix[i][j] = m[i][j] * x;
         }
     }
     return new_matrix;
 }
 
-Matrix operator*(const Matrix &m, Float x) {
+Matrix operator*(const Matrix &m, double x) {
     Matrix new_matrix(std::move(m.dimensions()));
 
-    for(UInt i = 0; i < m.dimensions().row_count; i++){
-        for(UInt j = 0; j < m.dimensions().column_count; j++){
+    for(unsigned int i = 0; i < m.dimensions().row_count; i++){
+        for(unsigned int j = 0; j < m.dimensions().column_count; j++){
             new_matrix[i][j] = m[i][j] * x;
         }
     }
     return new_matrix;
 }
 
-Matrix operator/(const Matrix &m, Float x) {
+Matrix operator/(const Matrix &m, double x) {
     Matrix new_matrix(std::move(m.dimensions()));
 
-    for(UInt i = 0; i < m.dimensions().row_count; i++){
-        for(UInt j = 0; j < m.dimensions().column_count; j++){
+    for(unsigned int i = 0; i < m.dimensions().row_count; i++){
+        for(unsigned int j = 0; j < m.dimensions().column_count; j++){
             new_matrix[i][j] = m[i][j] / x;
         }
     }
@@ -155,13 +155,13 @@ Matrix operator/(const Matrix &m, Float x) {
 Matrix operator*(const Matrix& m1, const Matrix& m2){
     if(m1.dimensions().column_count != m2.dimensions().row_count)
         throw std::invalid_argument("Matrix multiplication: Dimensions are not valid");
-    UInt n = m1.dimensions().column_count;
+    unsigned int n = m1.dimensions().column_count;
 
     Matrix m(m1.dimensions().row_count, m2.dimensions().column_count);
 
-    for(UInt i = 0; i < m1.dimensions().row_count; i++){ // 3
-        for(UInt j = 0; j < m2.dimensions().column_count; j++){ // 4
-            for(UInt k = 0; k < n; k++){
+    for(unsigned int i = 0; i < m1.dimensions().row_count; i++){ // 3
+        for(unsigned int j = 0; j < m2.dimensions().column_count; j++){ // 4
+            for(unsigned int k = 0; k < n; k++){
                 m[i][j] += m1[i][k] * m2[k][j];
             }
         }
@@ -176,8 +176,8 @@ Matrix operator+(const Matrix& m1, const Matrix& m2){
         throw std::invalid_argument("Matrix Addition: Dimensions are not valid");
     Matrix m(m1.dimensions().row_count, m1.dimensions().column_count);
 
-    for(UInt i = 0; i < m1.dimensions().row_count; i++){
-        for(UInt j = 0; j < m1.dimensions().column_count; j++){
+    for(unsigned int i = 0; i < m1.dimensions().row_count; i++){
+        for(unsigned int j = 0; j < m1.dimensions().column_count; j++){
             m[i][j] = m1[i][j] + m2[i][j];
         }
     }
