+++ sources/src/RTFEM/DataStructure/Vector3.cpp
@@ -3,7 +3,7 @@
 namespace rtfem{
 Vector3::Vector3() : x(0), y(0), z(0){}
 
-Vector3::Vector3(Float x, Float y, Float z) : x(x), y(y), z(z){}
+Vector3::Vector3(double x, double y, double z) : x(x), y(y), z(z){}
 
 Vector3::~Vector3(){}
