+++ sources/src/RTFEM/DataStructure/Vector4.cpp
@@ -3,7 +3,7 @@
 namespace rtfem{
 Vector4::Vector4() : x(0), y(0), z(0), w(0){}
 
-Vector4::Vector4(Float x, Float y, Float z, Float w) : x(x), y(y), z(z), w(w){}
+Vector4::Vector4(double x, double y, double z, double w) : x(x), y(y), z(z), w(w){}
 
 Vector4::~Vector4(){}
